linkedin-bot/requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
python-dotenv>=1.0
uvicorn>=0.27
fastapi>=0.109
jinja2>=3.1
aiofiles>=23.0
duckduckgo-search>=4.0
google-genai>=1.0
Pillow>=10.0
SpeechRecognition>=3.10
gTTS>=2.5
